fix: refactor botc
